+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 1. `activate-keypads` function.
+2. _erlang_ bindings.
 
 
 ## [0.8.5](https://github.com/uhppoted/uhppoted-codegen/releases/tag/v0.8.5) - 2023-06-13